requirements.txt
asgiref==3.7.2
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
iniconfig==2.0.0
packaging==23.1
Pillow==10.0.0
pluggy==1.3.0
psycopg2-binary==2.9.6
PyJWT==2.8.0
pytest==7.4.0
pytz==2023.3
sqlparse==0.4.4